Swap to uint64_t from uint32_t as the default positional value type

+++ FretFrameworkV2/Notes/SpecialPhrase.h
@@ -23,5 +23,5 @@ enum class SpecialPhraseType
 struct SpecialPhrase : public Sustained
 {
 	SpecialPhraseType type;
-	constexpr SpecialPhrase(SpecialPhraseType _type, uint32_t length) : type(_type) { setLength(length); }
+	constexpr SpecialPhrase(SpecialPhraseType _type, uint64_t length) : type(_type) { setLength(length); }
 };
