+++ FretFrameworkV2/Notes/Sustained.cpp
@@ -1,2 +1,2 @@
 #include "Sustained.h"
-uint32_t Sustained::s_minLength = 64;
+uint64_t Sustained::s_minLength = 64;
